[fri] Split binius_prover::fri into submodules (#102)
# [fri] Split binius_prover::fri into submodules

Split the FRI prover implementation into separate files for better organization:
- `commit.rs`: Code for encoding and committing messages
- `fold.rs`: FRI folding logic
- `query.rs`: Query phase implementation
- `tests.rs`: Comprehensive test suite

Also made `fold_interleaved` and related functions private since they're implementation details.

# [fri] Allow different data types for matrix and vector in fold operations

Modified `fold_cols` and `fold_rows` to accept different data types for the matrix and vector parameters, enabling more flexible usage patterns.

# [fri] Add NTT domain compatibility test

Added a test to verify that FRI folding and encoding operations commute correctly, ensuring the implementation works properly with NTT domains.
